Separate sleep timer intent duration
